test for tsconsole before creation, reorganized ts specific binaries

+++ recipes-core/initscripts/initscripts-1.0/tsinit.sh
@@ -9,6 +9,7 @@
 # Description: 
 ### END INIT INFO
 
-cp -P /initrd/dev/tsconsole /dev/tsconsole
+test -e /dev/tsconsole || cp -P /initrd/dev/tsconsole /dev/tsconsole
+
 rdev="$(mountpoint -d /)"
 mknod -m 600 /dev/rootdev b ${rdev%:*} ${rdev#*:}
